Updated the Entity Context for another table for saving the PayPal Customers.
Fixed the constants.

Updated the PayPalService to Create a customer in braintree. On Success, this writes to our DB.

Updated the Service to insert the PayPalCustomer into our DB.

Also added a GetCustomer method that get's the PayPal Customer.

Updated AccountController to Create a PayPalCustomer in Braintree as soon as the person registers.

We get the newly created CustomerId from our DB and create the clientToken with the CustomerId. This is used if the customer is already on file, it pre-populates the CC info with the card they have on file.